Port precision_from_str to Rust
