+++ nautilus_core/core/src/lib.rs
@@ -17,6 +17,7 @@ pub mod correctness;
 pub mod cvec;
 pub mod datetime;
 pub mod enums;
+pub mod parsing;
 pub mod string;
 pub mod time;
 pub mod uuid;
