+++ nautilus_trader/core/includes/core.h
@@ -93,6 +93,16 @@ const char *message_category_to_cstr(enum MessageCategory value);
 enum MessageCategory message_category_from_cstr(const char *ptr);
 
 /**
+ * Return the decimal precision inferred from the given C string.
+ *
+ * # Safety
+ * - Assumes `ptr` is a valid C string pointer.
+ * # Panics
+ * - If `ptr` is null.
+ */
+uint8_t precision_from_cstr(const char *ptr);
+
+/**
  * Drops the C string memory at the pointer.
  *
  * # Safety
